dev: refactor kakarot_rpc_core::test_utils to a kakarot_test_utils crate
