+++ crates/core/src/client/mod.rs
@@ -62,7 +62,7 @@ use crate::models::transaction_receipt::StarknetTransactionReceipt as Transactio
 use crate::models::ConversionError;
 
 pub struct KakarotClient<P: Provider + Send + Sync> {
-    starknet_provider: Arc<P>,
+    pub starknet_provider: Arc<P>,
     deployer_account: SingleOwnerAccount<Arc<P>, LocalWallet>,
     kakarot_contract: KakarotContract<P>,
     network: Network,
