+++ crates/core/src/lib.rs
@@ -3,4 +3,3 @@ pub mod client;
 pub mod contracts;
 pub mod mock;
 pub mod models;
-pub mod test_utils;
