+++ crates/eth-rpc/tests/integration.rs
@@ -5,8 +5,8 @@ mod integration_tests {
     use dotenv::dotenv;
     use ethers::prelude::{Block as EthersBlock, Http as EthersHttp, H256 as EthersH256};
     use kakarot_rpc::test_utils::start_kakarot_rpc_server;
-    use kakarot_rpc_core::test_utils::deploy_helpers::KakarotTestEnvironmentContext;
-    use kakarot_rpc_core::test_utils::fixtures::kakarot_test_env_ctx;
+    use kakarot_test_utils::deploy_helpers::KakarotTestEnvironmentContext;
+    use kakarot_test_utils::fixtures::kakarot_test_env_ctx;
     use reth_primitives::U64;
     use rstest::*;
