+++ crates/hive-utils/Cargo.toml
@@ -18,6 +18,8 @@ starknet_api = { workspace = true }
 tokio = { workspace = true }
 url = { workspace = true }
 
+kakarot_test_utils = { path = "../test-utils" }
+
 [dev-dependencies]
 cargo-husky = { workspace = true }
 katana-core = { workspace = true }
