+++ crates/test-utils/src/fixtures.rs
@@ -1,7 +1,7 @@
 use futures::executor::block_on;
 use rstest::*;
 
-use crate::test_utils::deploy_helpers::KakarotTestEnvironmentContext;
+use crate::deploy_helpers::KakarotTestEnvironmentContext;
 
 #[fixture]
 pub fn kakarot_test_env_ctx() -> KakarotTestEnvironmentContext {
